# Added fully unboxed interface to Genotype.