Use in expr.