@cseed 2017-02-22
Timing:

```
%%time
print(hc.read('profile225.vds')
      .filter_genotypes(
          '''
          g.dp > 400 ||
            (g.isHomRef && (g.ad[0] / g.dp < 0.9 || g.gq < 20)) ||
            (g.isHomVar && (g.ad[1] / g.dp < 0.9 || g.pl[0] < 20)) ||
            (g.isHet && ( (g.ad[0] + g.ad[1]) / g.dp < 0.9 || g.ad[1] / g.dp < 0.20 || g.pl[0] < 20 ))
          ''', keep=True)
 .count(genotypes=True))
```

master: 68s 66s 62s
this change: 57.1s 56.1s 60.3s