@cseed 2017-02-23
Addressed comments, back to you.